requirements.txt
bleak==0.22.3
paho-mqtt==2.1.0
PyYAML==6.0.2
victron-ble==0.9.2
pyserial==3.5
astral==3.2
